// src/api/generated/models/member-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 4차 프로젝트 7팀 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 회원가입 DTO
 * @export
 * @interface MemberDTO
 */
export interface MemberDTO {
    /**
     * 회원 ID
     * @type {number}
     * @memberof MemberDTO
     */
    'id'?: number;
    /**
     * 이름
     * @type {string}
     * @memberof MemberDTO
     */
    'name': string;
    /**
     * 사용자명
     * @type {string}
     * @memberof MemberDTO
     */
    'username': string;
    /**
     * 비밀번호
     * @type {string}
     * @memberof MemberDTO
     */
    'password': string;
    /**
     * 비밀번호 확인
     * @type {string}
     * @memberof MemberDTO
     */
    'password2': string;
    /**
     * 닉네임
     * @type {string}
     * @memberof MemberDTO
     */
    'nickname': string;
    /**
     * 이메일
     * @type {string}
     * @memberof MemberDTO
     */
    'email': string;
    /**
     * 전화번호
     * @type {string}
     * @memberof MemberDTO
     */
    'phoneNum': string;
    /**
     * 권한
     * @type {string}
     * @memberof MemberDTO
     */
    'role': string;
    /**
     * 주소
     * @type {string}
     * @memberof MemberDTO
     */
    'address': string;
}